requirements.txt
pyTelegramBotAPI==4.14.0
openai==1.12.0
notion-client==2.2.1
requests==2.31.0
python-dotenv==1.0.0
pytz==2024.1